Make it usable in a Xamarin.Forms project
