+++ README.md
@@ -1,27 +1,28 @@
 # XamarinAndroidFFmpeg (Xamarin.FFmpeg)
-FFmpeg library to run ffmeg commands over Xamarin.Android.
+FFmpeg library to run ffmeg commands over Xamarin.Forms
 
 ## About
-<ul>
-<li>Created to support a solution that needed ffmpeg to control camera devices.</li>
-</ul>
+* Created to support a solution that needed ffmpeg to control camera devices.
+* Edited to allow usage on Xamarin.Forms
 
 ## Big thanks
-<ul>
-<li>FFmpeg library adapted from https://github.com/neurospeech/xamarin-android-ffmpeg</li>
-<li>Original library in java https://github.com/WritingMinds/ffmpeg-android-java</li>
-</ul>
+* FFmpeg library adapted from https://github.com/neurospeech/xamarin-android-ffmpeg
+* Original library in java https://github.com/WritingMinds/ffmpeg-android-java
 
 ## License
-<ul>
-  <li>MIT</li>
-  <li>Besides that, to use this library you must accept the licensing terms mentioned in the source project at https://github.com/WritingMinds/ffmpeg-android-java</li>
-</ul>
+* MIT
+* Besides that, to use this library you must accept the licensing terms mentioned in the source project at https://github.com/WritingMinds/ffmpeg-android-java
 
 ## Nuget package
 You can download Xamarin.FFmpeg package from Nuget Package Manager or run following command in Nuget Package Console.
 ```
-Install-Package Xamarin.FFmpeg -Version 1.0.4
+Install-Package Xamarin.FFmpeg
 ```
 ## Usage
-On the way...
+
+### Android
+
+In MainActivity.cs before ``Forms.Init(this, bundle)``:
+```cs
+Xamarin.FFmpeg.Android.Init.Initialize(this.BaseContext);
+```
