+++ XamarinAndroidFFmpeg.sln
@@ -1,16 +1,16 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27130.2020
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29201.188
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Xamarin.FFmpeg.Nuget", "Xamarin.FFmpeg.Nuget\Xamarin.FFmpeg.Nuget.csproj", "{0451BAEF-DF2E-4B98-8644-94EE9415E389}"
 	ProjectSection(ProjectDependencies) = postProject
 		{1049745B-3D4C-4DFF-AD8B-DC46F460B5D4} = {1049745B-3D4C-4DFF-AD8B-DC46F460B5D4}
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Xamarin.FFmpeg", "Xamarin.FFmpeg\Xamarin.FFmpeg.csproj", "{1049745B-3D4C-4DFF-AD8B-DC46F460B5D4}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Xamarin.FFmpeg.Android", "Xamarin.FFmpeg.Android\Xamarin.FFmpeg.Android.csproj", "{1049745B-3D4C-4DFF-AD8B-DC46F460B5D4}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{A2C28ED0-0741-4CDE-B9FA-604C145DE4D9}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Xamarin.FFmpeg", "Xamarin.FFMpeg\Xamarin.FFmpeg.csproj", "{807F181E-43AA-4516-A86D-D68BB9EF5E95}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -26,6 +26,10 @@ Global
 		{1049745B-3D4C-4DFF-AD8B-DC46F460B5D4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1049745B-3D4C-4DFF-AD8B-DC46F460B5D4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1049745B-3D4C-4DFF-AD8B-DC46F460B5D4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{807F181E-43AA-4516-A86D-D68BB9EF5E95}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{807F181E-43AA-4516-A86D-D68BB9EF5E95}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{807F181E-43AA-4516-A86D-D68BB9EF5E95}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{807F181E-43AA-4516-A86D-D68BB9EF5E95}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
