--- src/Transforms/Metadata.xml
@@ -1,14 +1,0 @@
-<metadata>
-  <!--
-  This sample removes the class: android.support.v4.content.AsyncTaskLoader.LoadTask:
-  <remove-node path="/api/package[@name='android.support.v4.content']/class[@name='AsyncTaskLoader.LoadTask']" />
-  
-  This sample removes the method: android.support.v4.content.CursorLoader.loadInBackground:
-  <remove-node path="/api/package[@name='android.support.v4.content']/class[@name='CursorLoader']/method[@name='loadInBackground']" />
-  -->
-
-  <attr path="/api/package[@name='com.github.hiteshsondhi88.libffmpeg']/interface[@name='FFmpegInterface']" name="visibility">public</attr>
-  <attr path="/api/package[@name='com.github.hiteshsondhi88.libffmpeg']/interface[@name='ResponseHandler']" name="visibility">public</attr>
-
-
-</metadata>
